+ RFE: allow SO to set the weightclass of a unit based on build table settings + RFE: allow force size to face to use float

+++ trunk/src/common/Army.java
@@ -47,7 +47,7 @@ public class Army {
 	
 	private boolean armyPlayerLocked = false; //Used by players to keep armies from being cleared
 
-	private int opForceSize = NO_LIMIT;
+	private float opForceSize = NO_LIMIT;
 	
 	private Hashtable<Integer,Integer> c3Network = new Hashtable<Integer,Integer>();
 
@@ -316,11 +316,11 @@ public class Army {
 		return Math.max(1, count);
 	}
 	
-	public int getOpForceSize() {
+	public float getOpForceSize() {
 		return this.opForceSize;
 	}
 	
-	public void setOpForceSize(int force) {
+	public void setOpForceSize(float force) {
 		this.opForceSize = force;
 	}
